# Conflicts:
#	src/main/java/org/prebid/server/auction/BidResponseCreator.java
#	src/main/java/org/prebid/server/auction/ExchangeService.java
#	src/main/java/org/prebid/server/bidder/sovrn/SovrnBidder.java
#	src/main/java/org/prebid/server/deals/targeting/RequestContext.java
#	src/main/java/org/prebid/server/metric/MetricName.java
#	src/main/java/org/prebid/server/metric/Metrics.java
#	src/main/resources/application.yaml
#	src/test/java/org/prebid/server/bidder/sovrn/SovrnBidderTest.java
#	src/test/java/org/prebid/server/floors/BasicPriceFloorAdjusterTest.java
#	src/test/java/org/prebid/server/floors/PriceFloorFetcherTest.java

+++ src/test/groovy/org/prebid/server/functional/model/request/auction/Deal.groovy
@@ -9,7 +9,7 @@ import groovy.transform.ToString
 class Deal {
 
     String id
-    Float bidFloor
+    BigDecimal bidFloor
     String bidFloorCur
     Integer at
     List<String> wseat
